+++ CHANGELOG.md
@@ -10,6 +10,11 @@ All changes to this project will be documented in this file.
 > - 🏠 [Internal]
 > - 💅 [Polish]
 
+## Unreleased
+
+#### 🚀 New Feature
+- allow using snapshot-content object for `remap-snapshot`
+
 ## v20.5.0 (2017-06-30)
 
 #### 🚀 New Feature